+++ src/OpenSleigh.Core/ExceptionPolicies/DelayFactory.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace OpenSleigh.Core.ExceptionPolicies
+{
+    public delegate TimeSpan DelayFactory(int executionIndex);
+}
