+++ src/OpenSleigh.Core/ExceptionPolicies/ExceptionContext.cs
@@ -4,5 +4,5 @@ using System.Diagnostics.CodeAnalysis;
 namespace OpenSleigh.Core.ExceptionPolicies
 {
     [ExcludeFromCodeCoverage]
-    internal record ExceptionContext(Exception Exception, int ExecutionIndex);
+    public record ExceptionContext(Exception Exception, int ExecutionIndex);
 }
