+++ src/OpenSleigh.Core/ExceptionPolicies/Policy.cs
@@ -8,14 +8,7 @@ namespace OpenSleigh.Core.ExceptionPolicies
         {
             var builder = new RetryPolicyBuilder();
             builderAction(builder);
-            return new RetryPolicy(builder.MaxRetries, builder.ExceptionFilters, builder.OnExceptionHandler);
-        }
-
-        public static PolicyBase DelayedRetry(Action<DelayedRetryPolicyBuilder> builderAction)
-        {
-            var builder = new DelayedRetryPolicyBuilder();
-            builderAction(builder);
-            return new DelayedRetryPolicy(builder.MaxRetries, builder.ExceptionFilters, builder.DelayFactory, builder.OnExceptionHandler);
+            return new RetryPolicy(builder.MaxRetries, builder.ExceptionFilters, builder.DelayFactory, builder.OnExceptionHandler);
         }
     }
 }
